pll draft complete save for table/scale selection and LED dispaly

+++ osc1_rev5/Src/main_sample_generation.c
@@ -251,7 +251,7 @@ static inline int getSampleQuinticSpline(int phase, uint32_t morph, int * delta)
 			))
 		));
 
-	int deltaSign = ((sample3 - sample2) >> 31);
+	int deltaSign = ((uint32_t)(sample3 - sample2) >> 31);
 
 	*delta = deltaSign;
